Create 'normalizeMRZString()' as a module import.
